fix bad refactoring
