+++ batch/batch/cloud/worker.py
@@ -1,8 +1,5 @@
-from typing import TYPE_CHECKING, Dict
+from typing import Dict
 
-from gear.cloud_config import get_global_config
-
-from .gcp.driver.driver import GCPDriver
 from .gcp.worker.credentials import GCPUserCredentials
 from .gcp.worker.disk import GCPDisk
 from .gcp.instance_config import GCPInstanceConfig
@@ -10,18 +7,6 @@ from .gcp.instance_config import GCPInstanceConfig
 from ..worker.credentials import CloudUserCredentials
 from ..worker.disk import CloudDisk
 from ..instance_config import InstanceConfig
-
-
-if TYPE_CHECKING:
-    from ..inst_coll_config import InstanceCollectionConfigs  # pylint: disable=cyclic-import
-    from ..driver.driver import CloudDriver  # pylint: disable=cyclic-import
-
-
-async def get_cloud_driver(app, machine_name_prefix: str, namespace: str, inst_coll_configs: 'InstanceCollectionConfigs',
-                           credentials_file: str) -> 'CloudDriver':
-    cloud = get_global_config()['cloud']
-    assert cloud == 'gcp'
-    return await GCPDriver.create(app, machine_name_prefix, namespace, inst_coll_configs, credentials_file)
 
 
 def get_cloud_disk(instance_name: str,
